More refactoring, cleanup.
