+++ .vscode/c_cpp_properties.json
@@ -3,7 +3,7 @@
         {
             "name": "Linux",
             "includePath": [
-                "${workspaceFolder}/src/libs/",
+                "${workspaceFolder}/src/utils/",
                 "${workspaceFolder}/src/btservice/",
                 "/usr/include/glib-2.0/",
                 "/usr/include/dbus-1.0",
