+++ src/Makefile.am
@@ -1,3 +1,3 @@
-SUBDIRS = btservice stepper jscturret
+SUBDIRS = utils btservice jscturret stepper
 EXTRA_DIST = autogen.sh
 ACLOCAL_AMFLAGS = -I m4
