+++ src/btservice/btconstant.h
@@ -1,0 +1,19 @@
+#define BLUETOOTH_BASE_128_UUID "00000000-0000-1000-8000-00805F9B34FB"
+#define SERIAL_PORT_PROFILE_UUID "0000110e-0000-1000-8000-00805f9b34fb"
+
+#define BLUETOOTH_NAME_MAX_LENGTH_BYTES 248
+#define BLUETOOTH_ADDRESS_STRING_SIZE 18
+#define BLUETOOTH_UUID_STRING_SIZE 64
+#define BLUETOOTH_DISCOVERY_MAX_WAIT_SECONDS 60
+
+#define ERR_CANNOT_GET_DBUS_CONNECTION 500
+#define ERR_CANNOT_ENABLE_ADAPTER 501
+#define ERR_CANNOT_DISABLE_ADAPTER 502
+#define ERR_CANNOT_START_SCAN 503
+#define ERR_CANNOT_STOP_SCAN 504
+#define ERR_CUSTOM_PATH_INVALID 505
+#define ERR_CANNOT_REGISTER_PROFILE 506
+#define ERR_CANNOT_CONNECT_DEVICE 507
+#define ERR_CANNOT_CALL_METHOD 508
+
+#define RESULT_OK 200
