+++ src/btservice/btserver.h
@@ -1,0 +1,1 @@
+int register_service(char *service_path, char *service_name, int service_channel, char *service_uuid);
