+++ src/buildme.sh
@@ -3,3 +3,8 @@
 mkdir -p build
 cd build
 ../configure --prefix=${PWD}/build/ && make
+
+if [ "$1" = "install" ];
+then
+  make install
+fi
