+++ src/utils/Makefile.am
@@ -1,0 +1,8 @@
+lib_LTLIBRARIES = libjscutils.la
+libjscutils_la_SOURCES = logger.c
+include_HEADERS = logger.h
+libjscutils_la_CPPFLAGS = -I/usr/include/glib-2.0 -I/usr/lib/x86_64-linux-gnu/glib-2.0/include
+libjscutils_la_CPPFLAGS += -I/usr/include/glib-2.0 -I/usr/lib/arm-linux-gnueabihf/glib-2.0/include
+libjscutils_la_CPPFLAGS += -pthread -I/usr/include/libmount -I/usr/include/blkid
+libjscutils_la_CPPFLAGS += -I/usr/include/gio-unix-2.0
+libjscutils_la_LIBADD = -lglib-2.0 -lgobject-2.0
