+++ src/utils/logger.h
@@ -1,0 +1,3 @@
+void _info(char *message, int crlf);
+void info(char *message);
+void ninfo(char *message);
